mazal mazal

+++ minitalk/client.c
@@ -12,6 +12,16 @@
 
 #include "minitalk.h"
 
+volatile sig_atomic_t ack_received = 0;
+
+void handle_ack(int signal, siginfo_t *info, void *context)
+{
+	(void)info;
+	(void)context;
+	(void)signal;
+	ack_received = 1;
+}
+
 void	send_signal(int pid, char c)
 {
 	int	i;
@@ -37,13 +47,14 @@ void	send_signal(int pid, char c)
 				exit(1);
 			}
 		}
-		usleep(100);
+		usleep(500);
 		i--;
 	}
 }
 
 int main(int argc, char *argv[])
 {
+	struct sigaction sa;
 	int	server_pid;
 	int	i;
 	char	*message;
@@ -56,11 +67,23 @@ int main(int argc, char *argv[])
 	server_pid = atoi(argv[1]);
 	message = argv[2];
 	i = 0;
+	sigemptyset(&sa.sa_mask);
+	sa.sa_sigaction = handle_ack;
+	sa.sa_flags = SA_SIGINFO;
+	if (sigaction(SIGUSR1, &sa, NULL) == -1)
+	{
+		printf("Error regestering SIGUSR1 Handler");
+		exit(1);
+	}
 	while (message[i] != '\0')
 	{
 		send_signal(server_pid, message[i]);
 		i++;
+		while (!ack_received)
+			pause();
+		ack_received = 0;
 	}
 	send_signal(server_pid, '\0');
+	printf("Server received the message Perfectly 🚀");
 	return (0);
 }
